allow persistent dbs
while creating new dbs for each connection may be the desired to use
the server, it's hard to debug this server on its own.

If provided HTTP_PERSISTENT_DB or WS_PERSISTENT_DB, those files will
be used and they won't be deleted.

This is mostly used to develop and debug the test servers themselves.